refactor: access countをcomponentに切り出し
