+++ src/components/access-count/index.ts
@@ -1,0 +1,1 @@
+export { AccessCount } from './access-count'
