+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { MouseStalker } from './mouse-stalker'
 export { BgBody } from './background'
 export { MineSweeper } from './mine-sweeper'
+export { AccessCount } from './access-count'
